+++ src/lib/packages/base/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @cluue/base
 
+## 1.0.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @cluue/icons@1.0.23
+
 ## 1.0.25
 
 ### Patch Changes