+++ src/lib/packages/forms/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cluue/forms
 
+## 1.0.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @cluue/icons@1.0.23
+  - @cluue/base@1.0.26
+
 ## 1.0.25
 
 ### Patch Changes