+++ src/lib/packages/icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cluue/icons
 
+## 1.0.23
+
+### Patch Changes
+
+- Fix: drift cells
+
 ## 1.0.22
 
 ### Patch Changes